# docs/api/ai-hub-openapi.yaml
openapi: 3.0.3
info:
  title: AI Hub Platform API
  description: |
    企业级AI应用平台API文档

    ## 🎯 平台概述
    AI Hub是一个企业级AI应用平台，集成了多种AI模型，提供流式响应、成本跟踪和多模���AI处理能力。

    ## 🚀 Week 7 新功能
    - **多模态AI服务**: 图像识别、语音处理、文档理解
    - **智能工作流引擎**: 可视化工作流设计与执行
    - **API文档与测试**: 完整的API文档和交互式测试

    ## 🔗 主要功能
    - AI聊天和对话管理
    - 多模态AI处理（图像、语音、文档）
    - 智能工作流引擎
    - 企业级用户管理
    - API密钥和配额管理
    - 实时监控和审计

  version: 1.0.0
  contact:
    name: AI Hub Team
    email: support@ai-hub.com
    url: https://ai-hub.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8001
    description: 开发环境
  - url: https://api.ai-hub.com
    description: 生产环境
...
paths:
...
  /api/v1/multimodal/vision/analyze:
    post:
      tags:
        - Multi-Modal AI
      summary: 分析图像内容
      description: |
        使用AI分析图像内容，支持多种分析类型：
        - general: 通用图像分析
        - ocr: 文字识别(OCR)
        - object_detection: 物体检测
        - classification: 图像分类
        - sentiment: 情感分析
        - content_moderation: 内容审核
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - image
              properties:
                image:
                  type: string
                  format: binary
                  description: 要分析的图像文件
                analysis_type:
                  type: string
                  enum: [general, ocr, object_detection, classification, sentiment, content_moderation]
                  default: general
                  description: 分析类型
                prompt:
                  type: string
                  default: "请详细分析这张图片的内容"
                  description: 分析提示词
      responses:
        '200':
          description: 分析成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageAnalysisResponse'
        '400':
          description: 请求参数错误
        '500':
          description: 服务器内部错误
  /api/v1/multimodal/speech/transcribe:
    post:
      tags:
        - Multi-Modal AI
      summary: 音频转文字
      description: 将音频文件转换为文字，支持多种任务类型
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - audio
              properties:
                audio:
                  type: string
                  format: binary
                  description: 要转录的音频文件
                task_type:
                  type: string
                  enum: [transcription, translation, dictation, command_recognition]
                  default: transcription
                  description: 任务类型
                language:
                  type: string
                  default: auto
                  description: 语言代码
      responses:
        '200':
          description: 转录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpeechTranscriptionResponse'
        '400':
          description: 请求参数错误
        '500':
          description: 服务器内部错误
  /api/v1/multimodal/speech/synthesize:
    post:
      tags:
        - Multi-Modal AI
      summary: 文字转语音
      description: 将文本转换为语音文件
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - text
              properties:
                text:
                  type: string
                  description: 要转换为语音的文本
                voice:
                  type: string
                  enum: [alloy, echo, fable, onyx, nova, shimmer]
                  default: alloy
                  description: 声音类型
                language:
                  type: string
                  default: en
                  description: 语言代码
      responses:
        '200':
          description: 语音合成成功
          content:
            audio/mpeg:
              schema:
                type: string
                format: binary
        '400':
          description: 请求参数错误
        '500':
          description: 服务器内部错误
  /api/v1/multimodal/document/analyze:
    post:
      tags:
        - Multi-Modal AI
      summary: 文档内容分析
      description: 智能分析文档内容，提取关键信息并生成摘要
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - document
              properties:
                document:
                  type: string
                  format: binary
                  description: 要分析的文档文件
                query:
                  type: string
                  default: "请总结这个文档的主要内容"
                  description: 分析查询
      responses:
        '200':
          description: 分析成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentAnalysisResponse'
        '400':
          description: 请求参数错误
        '500':
          description: 服务器内部错误
  /api/v1/multimodal/process:
    post:
      tags:
        - Multi-Modal AI
      summary: 多模态综合处理
      description: 同时处理多种类型的输入内容
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
                  description: 图像文件
                audio:
                  type: string
                  format: binary
                  description: 音频文件
                document:
                  type: string
                  format: binary
                  description: 文档文件
                text:
                  type: string
                  description: 文本内容
                task:
                  type: string
                  default: analyze
                  description: 处理任务
                prompt:
                  type: string
                  description: 通用提示词
      responses:
        '200':
          description: 处理成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MultiModalProcessResponse'
        '400':
          description: 请求参数错误
        '500':
          description: 服务器内部错误
...
components:
  schemas:
...
    ImageAnalysisResponse:
      type: object
      properties:
        success:
          type: boolean
          description: 是否成功
        description:
          type: string
          description: 分析描述
        tags:
          type: array
          items:
            type: string
          description: 标签列表
        confidence:
          type: number
          format: float
          description: 置信度
        objects:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
                description: 对象名称
              confidence:
                type: number
                format: float
                description: 置信度
          description: 识别的对象
        text_content:
          type: string
          description: 识别的文字内容
        sentiment:
          type: string
          description: 情感分析结果
        moderation_flags:
          type: array
          items:
            type: string
          description: 内容审核标志
        metadata:
          type: object
          description: 元数据
        processing_time:
          type: number
          format: float
          description: 处理时间(秒)
        created_at:
          type: string
          format: date-time
          description: 创建时间
    SpeechTranscriptionResponse:
      type: object
      properties:
        success:
          type: boolean
          description: 是否成功
        text:
          type: string
          description: 转录的文本
        confidence:
          type: number
          format: float
          description: 置信度
        language:
          type: string
          description: 语言代码
        duration:
          type: number
          format: float
          description: 音频时长(秒)
        segments:
          type: array
          items:
            type: object
            properties:
              start:
                type: number
                description: 开始时间
              end:
                type: number
                description: 结束时间
              text:
                type: string
                description: 文本片段
              confidence:
                type: number
                description: 置信度
          description: 转录片段
        translated_text:
          type: string
          description: 翻译文本
        metadata:
          type: object
          description: 元数据
        processing_time:
          type: number
          format: float
          description: 处理时间(秒)
        created_at:
          type: string
          format: date-time
          description: 创建时间
    DocumentAnalysisResponse:
      type: object
      properties:
        success:
          type: boolean
          description: 是否成功
        content:
          type: string
          description: 文档内容
        summary:
          type: string
          description: 文档摘要
        key_points:
          type: array
          items:
            type: string
          description: 关键要点
        tables:
          type: array
          items:
            type: object
          description: 提取的表格
        images:
          type: array
          items:
            type: object
          description: 提取的图片
        metadata:
          type: object
          description: 元数据
        processing_time:
          type: number
          format: float
          description: 处理时间(秒)
        created_at:
          type: string
          format: date-time
          description: 创建时间
    MultiModalProcessResponse:
      type: object
      properties:
        success:
          type: boolean
          description: 是否成功
        task:
          type: string
          description: 处理任务
        results:
          type: object
          description: 处理结果
        metadata:
          type: object
          description: 元数据
        processing_time:
          type: number
          format: float
          description: 处理时间(秒)
        created_at:
          type: string
          format: date-time
          description: 创建时间
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT认证
...